add logic to the lib folder
